criar atributos para carroCondi
