+++ src/model/entities/carros/AlfaRomeo.java
@@ -1,0 +1,14 @@
+package model.entities.carros;
+
+public class AlfaRomeo extends carrosCondi {
+
+    public AlfaRomeo(String model) {
+        super(model);
+    }
+
+    @Override
+    public String toString() {
+        return "AlfaRomeo [modelo=" + getName() + "]";
+    }
+    
+}
