--- src/model/entities/carros/marcas.java
@@ -1,22 +1,0 @@
-package model.entities.carros;
-
-public class marcas {
-
-    private String modelo;
-
-    
-
-    public marcas(String model) {
-        this.modelo = model;
-    }
-
-    public String getName() {
-        return modelo;
-    }
-
-    public void setName(String name) {
-        this.modelo = name;
-    }
-
-    
-}
